requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
supabase>=2.6.0
pyrogram>=2.0.106
tgcrypto>=1.2.5
python-dotenv>=1.0.1
pydantic>=2.7.0
aiogram>=3.0.0
python-dotenv>=1.0.0

